pyproject.toml
[project]
name = "agents-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepeval>=3.8.8",
    "fastapi>=0.132.0",
    "google-genai>=1.64.0",
    "langchain-core>=1.2.15",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "langserve>=0.3.3",
    "langsmith>=0.7.6",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "ollama>=0.6.1",
    "pandas<3",
    "psycopg>=3.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rich<14",
    "seaborn>=0.13.2",
    "streamlit>=1.35.0",
    "supabase>=2.28.0",
    "tavily-python>=0.7.21",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "uvicorn>=0.41.0",
]
